Boom

+++ database/migrations/2020_03_17_004132_create_replies_table.php
@@ -16,8 +16,8 @@ class CreateRepliesTable extends Migration
         Schema::create('replies', function (Blueprint $table) {
             $table->id();
             $table->text('text');
-            $table->integer('question_id')->unsigned();
-            $table->integer('user_id');
+            $table->bigInteger('question_id')->unsigned();
+            $table->bigInteger('user_id');
             $table->foreign('question_id')->references('id')->on('questions')->onDelete('cascade');
